Revert "Fixing NoClassDefFoundError (#366)"
This reverts commit c912501a25e5e9c004c8ce6a27802f61df0d1cd7.
