+++ aws-xray-recorder-sdk-core/src/main/java/com/amazonaws/xray/entities/SubsegmentImpl.java
@@ -15,7 +15,6 @@
 
 package com.amazonaws.xray.entities;
 
-import com.amazonaws.xray.AWSXRayObjectMapper;
 import com.amazonaws.xray.AWSXRayRecorder;
 import com.amazonaws.xray.internal.SamplingStrategyOverride;
 import com.fasterxml.jackson.annotation.JsonIgnore;
@@ -146,7 +145,7 @@ public class SubsegmentImpl extends EntityImpl implements Subsegment {
     }
 
     private ObjectNode getStreamSerializeObjectNode() {
-        ObjectNode obj = (ObjectNode) AWSXRayObjectMapper.getInstance().valueToTree(this);
+        ObjectNode obj = (ObjectNode) mapper.valueToTree(this);
         obj.put("type", "subsegment");
         obj.put("parent_id", getParent().getId());
         obj.put("trace_id", parentSegment.getTraceId().toString());
@@ -157,7 +156,7 @@ public class SubsegmentImpl extends EntityImpl implements Subsegment {
     public String streamSerialize() {
         String ret = "";
         try {
-            ret = AWSXRayObjectMapper.getInstance().writeValueAsString(getStreamSerializeObjectNode());
+            ret = mapper.writeValueAsString(getStreamSerializeObjectNode());
         } catch (JsonProcessingException jpe) {
             logger.error("Exception while serializing entity.", jpe);
         }
@@ -168,8 +167,7 @@ public class SubsegmentImpl extends EntityImpl implements Subsegment {
     public String prettyStreamSerialize() {
         String ret = "";
         try {
-            ret = AWSXRayObjectMapper.getInstance().writerWithDefaultPrettyPrinter()
-                    .writeValueAsString(getStreamSerializeObjectNode());
+            ret = mapper.writerWithDefaultPrettyPrinter().writeValueAsString(getStreamSerializeObjectNode());
         } catch (JsonProcessingException jpe) {
             logger.error("Exception while serializing entity.", jpe);
         }
